replace target_io with loader_io in emmc code

+++ src/flash/emmc/imp.h
@@ -19,7 +19,7 @@ void emmc_device_add(struct emmc_device *c);
 int emmc_probe(struct emmc_device *emmc);
 
 int emmc_write_data_block(struct emmc_device *emmc, uint32_t *buffer, uint32_t address);
-int emmc_write_image(struct emmc_device *emmc, uint32_t *buffer, uint32_t address, int size);
+int emmc_write_image(struct emmc_device *emmc, uint8_t *buffer, uint32_t address, int size);
 int emmc_read_data_block(struct emmc_device *emmc, uint32_t *buffer, uint32_t address);
 int emmc_verify_image(struct emmc_device *emmc, uint8_t *buffer, uint32_t addr, int size);
